Fix/enhance THP integration.
Detect whether chunks start off as THP-capable by default (according to
the state of /sys/kernel/mm/transparent_hugepage/enabled), and use this
as the basis for whether to call pages_nohuge() once per chunk during
first purge of any of the chunk's page runs.

Add the --disable-thp configure option, as well as the the opt.thp
mallctl.

This resolves #541.

Change-Id: I3950580f76b6303d0c0afcb70027d76e214cb572
Signed-off-by: Alex Naidis <alex.naidis@linux.com>